refactor(calendar): extract DayOverflowPopover into its own file
Moves the popover subcomponent, shared EVENT_COLOR_CLASSES table, and
MAX_INLINE_EVENTS constant out of SimpleCalendar.tsx into a dedicated
DayOverflowPopover.tsx. SimpleCalendar now imports what it needs for
inline day cells and the overflow popover, and is ~100 lines shorter.

No behavior change — component tests and E2E are unchanged. This is
context segmentation for readability.
